# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Tea Sequence API
  version: 1.0.0
  description: >-
    Stable v1 contract for Tea Sequence. The initial configuration response
    explicitly describes non-production demonstration mode.
servers:
  - url: http://127.0.0.1:8000
paths:
...
  /api/v1/teas:
    get:
      operationId: listTeas
      summary: Search published teas and tea items
      parameters:
        - {name: q, in: query, schema: {type: string, maxLength: 100}}
        - {name: category, in: query, schema: {type: string}}
        - {name: entity_type, in: query, schema: {type: string, enum: [tea, tea_item]}}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
      responses:
        '200': {$ref: '#/components/responses/ApiSuccess'}
        '422': {$ref: '#/components/responses/ApiFailure'}
  /api/v1/teas/{tea_id}:
    get:
      operationId: getTea
      parameters: [{$ref: '#/components/parameters/TeaId'}]
      responses:
        '200': {$ref: '#/components/responses/ApiSuccess'}
        '404': {$ref: '#/components/responses/ApiFailure'}
  /api/v1/tea-items:
    get:
      operationId: listTeaItems
      parameters:
        - {name: tea_id, in: query, schema: {type: string, format: uuid}}
        - {name: sku, in: query, schema: {type: string, maxLength: 64}}
        - {name: batch_code, in: query, schema: {type: string, maxLength: 64}}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
      responses:
        '200': {$ref: '#/components/responses/ApiSuccess'}
  /api/v1/tea-items/{tea_item_id}:
    get:
      operationId: getTeaItem
      parameters: [{$ref: '#/components/parameters/TeaItemId'}]
      responses:
        '200': {$ref: '#/components/responses/ApiSuccess'}
        '404': {$ref: '#/components/responses/ApiFailure'}
  /api/v1/teas/{tea_id}/effects:
    get:
      operationId: getTeaEffects
      parameters:
        - {$ref: '#/components/parameters/TeaId'}
        - {name: tea_item_id, in: query, schema: {type: string, format: uuid}}
      responses:
        '200': {$ref: '#/components/responses/ApiSuccess'}
        '422': {$ref: '#/components/responses/ApiFailure'}
  /api/v1/teas/{tea_id}/brewing:
    get:
      operationId: getTeaBrewing
      parameters:
        - {$ref: '#/components/parameters/TeaId'}
        - {name: tea_item_id, in: query, schema: {type: string, format: uuid}}
      responses:
        '200': {$ref: '#/components/responses/ApiSuccess'}
        '422': {$ref: '#/components/responses/ApiFailure'}
  /api/v1/teas/{tea_id}/supplies:
    get:
      operationId: listTeaSupplies
      parameters:
        - {$ref: '#/components/parameters/TeaId'}
        - {name: tea_item_id, in: query, schema: {type: string, format: uuid}}
        - {$ref: '#/components/parameters/Page'}
        - {$ref: '#/components/parameters/PageSize'}
      responses:
        '200': {$ref: '#/components/responses/ApiSuccess'}
  /api/v1/sources/{source_id}:
    get:
      operationId: getPublicSource
      parameters: [{name: source_id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        '200': {$ref: '#/components/responses/ApiSuccess'}
        '404': {$ref: '#/components/responses/ApiFailure'}
  /api/v1/files/{file_id}/content:
    get:
      operationId: getPublicFile
      parameters: [{name: file_id, in: path, required: true, schema: {type: string, format: uuid}}]
      responses:
        '200': {description: Authorized binary file}
        '503': {$ref: '#/components/responses/ApiFailure'}
  /api/v1/questions:
    post:
      operationId: createQuestion
      security: []
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/QuestionCreate'}}}
      responses:
        '200': {$ref: '#/components/responses/ApiSuccess'}
        '409': {$ref: '#/components/responses/ApiFailure'}
        '422': {$ref: '#/components/responses/ApiFailure'}
  /api/v1/inquiries:
    post:
      operationId: createInquiry
      security: []
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/InquiryCreate'}}}
      responses:
        '201': {$ref: '#/components/responses/ApiSuccess'}
        '409': {$ref: '#/components/responses/ApiFailure'}
        '422': {$ref: '#/components/responses/ApiFailure'}
  /api/v1/feedback:
    post:
      operationId: createFeedback
      security: []
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/FeedbackCreate'}}}}
      responses: {'201': {$ref: '#/components/responses/ApiSuccess'}}
  /api/v1/events:
    post:
      operationId: createEvents
      security: []
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/EventBatchCreate'}}}}
      responses: {'200': {$ref: '#/components/responses/ApiSuccess'}}
...
components:
  securitySchemes:
    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: tea_sequence_session
      description: HttpOnly administrator session cookie for later admin operations.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Required by later state-changing operations; must be a UUID.
      schema:
        type: string
        format: uuid
...
    Page:
      name: page
      in: query
      schema: {type: integer, minimum: 1, default: 1}
    PageSize:
      name: page_size
      in: query
      schema: {type: integer, minimum: 1, maximum: 100, default: 20}
    TeaId:
      name: tea_id
      in: path
      required: true
      schema: {type: string, format: uuid}
    TeaItemId:
      name: tea_item_id
      in: path
      required: true
      schema: {type: string, format: uuid}
...
  responses:
    ApiSuccess:
      description: Standard success envelope; operation-specific data is defined by the prose contract.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/SuccessObject'}
    ApiFailure:
      description: Standard failure envelope; clients branch on error.code.
      content:
        application/json:
          schema: {$ref: '#/components/schemas/ErrorResponse'}
...
  schemas:
    SuccessObject:
      type: object
      additionalProperties: false
      required: [data, meta]
      properties:
        data: {type: object}
        meta: {$ref: '#/components/schemas/ResponseMeta'}
...
    QuestionCreate:
      type: object
      additionalProperties: false
      required: [question]
      properties:
        question: {type: string, minLength: 1, maxLength: 1000}
        intent: {type: [string, 'null'], enum: [general, brewing, null]}
        tea_id: {type: [string, 'null'], format: uuid}
        tea_item_id: {type: [string, 'null'], format: uuid}
    InquiryCreate:
      type: object
      additionalProperties: false
      required: [kind, need, contact, consent]
      properties:
        kind: {type: string, enum: [consultation, sample]}
        tea_id: {type: [string, 'null'], format: uuid}
        tea_item_id: {type: [string, 'null'], format: uuid}
        supply_offer_id: {type: [string, 'null'], format: uuid}
        need: {type: string, minLength: 1, maxLength: 1000}
        contact:
          type: object
          additionalProperties: false
          required: [channel, value]
          properties: {channel: {type: string, enum: [phone, email]}, value: {type: string, maxLength: 254}}
        consent:
          type: object
          additionalProperties: false
          required: [accepted, notice_version, purpose]
          properties: {accepted: {const: true}, notice_version: {type: string}, purpose: {const: inquiry_followup}}
    FeedbackCreate:
      type: object
      additionalProperties: false
      required: [target_type, target_id, rating]
      properties:
        target_type: {type: string, enum: [effect, brewing, answer]}
        target_id: {type: string, format: uuid}
        target_version: {type: [integer, 'null'], minimum: 1}
        rating: {type: string, enum: [helpful, unhelpful]}
        reason: {type: [string, 'null'], maxLength: 500}
    EventBatchCreate:
      type: object
      additionalProperties: false
      required: [events]
      properties:
        events: {type: array, minItems: 1, maxItems: 20, items: {type: object}}
...
    ResponseMeta:
      type: object
      additionalProperties: false
      required: [request_id, server_time]
      properties:
        request_id:
          type: string
          format: uuid
        server_time:
          type: string
          format: date-time
    ErrorDetail:
      type: object
      additionalProperties: true
      properties:
        field:
          type: string
        reason:
          type: string
    ApiErrorBody:
      type: object
      additionalProperties: false
      required: [code, message, details]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error, meta]
      properties:
        error:
          $ref: '#/components/schemas/ApiErrorBody'
        meta:
          $ref: '#/components/schemas/ResponseMeta'